add base class and rename
