+++ packages/scenes/src/index.ts
@@ -28,9 +28,12 @@ export { SceneTimeRange } from './core/SceneTimeRange';
 export { SceneTimeZoneOverride } from './core/SceneTimeZoneOverride';
 
 export { SceneQueryRunner, type QueryRunnerState } from './querying/SceneQueryRunner';
-export { SceneDataLayers } from './querying/SceneDataLayers';
+export { SceneDataLayerSet, SceneDataLayerSetBase } from './querying/SceneDataLayerSet';
 export { SceneDataLayerBase } from './querying/layers/SceneDataLayerBase';
-export { SceneDataLayerControls, SceneDataLayerControl } from './querying/layers/SceneDataLayerControls';
+export {
+  SceneDataLayerControls,
+  DataLayerControlSwitch as SceneDataLayerControl,
+} from './querying/layers/SceneDataLayerControls';
 export { SceneDataTransformer } from './querying/SceneDataTransformer';
 export { registerQueryWithController } from './querying/registerQueryWithController';
 export { registerRuntimeDataSource, RuntimeDataSource } from './querying/RuntimeDataSource';
